Move keyterm_to_int into the keyterm module
