+++ regparser/grammar/amdpar.py
@@ -9,8 +9,8 @@ from pyparsing import (
 
 from regparser.grammar import atomic, tokens, unified
 from regparser.grammar.utils import Marker, WordBoundaries
+from regparser.layer.key_terms import keyterm_to_int
 from regparser.tree.paragraph import p_levels
-from regparser.tree.xml_parser.paragraph_processor import ParagraphProcessor
 
 
 intro_text_marker = (
@@ -426,8 +426,7 @@ def tokenize_override_ps(match):
 _keyterm_label_part = (
     Suppress(Marker("keyterm"))
     + QuotedString(quoteChar='(', endQuoteChar=')')
-).setParseAction(
-    lambda m: "p{}".format(ParagraphProcessor.keyterm_to_int(m[0])))
+).setParseAction(lambda m: "p{}".format(keyterm_to_int(m[0])))
 _simple_label_part = Word(string.ascii_lowercase + string.ascii_uppercase
                           + string.digits)
 _label_part = _keyterm_label_part | _simple_label_part
