+++ tests/tree_xml_parser_paragraph_processor_tests.py
@@ -111,12 +111,6 @@ class ParagraphProcessorTest(TestCase):
         self.assertEqual(a.children[0].label, ['root', 'p1', 'a', 'p1'])
         self.assertEqual(b.label, ['root', 'p1', 'b'])
 
-    def test_keyterm_to_int(self):
-        """keyterm_to_int should standardize the keyterm"""
-        to_int = paragraph_processor.ParagraphProcessor.keyterm_to_int
-        self.assertEqual(to_int('Abc 123 More.'), to_int(' abc123 mOrE'))
-        self.assertTrue(to_int('a term') > 10000)
-
     def test_separate_intro_empty_nodes(self):
         """ Make sure separate_intro can handle an empty node list. """
         nodes = []
